requirements.txt
load_dotenv
openai 
unidecode
streamlit